refactor: refactored the RuleSettings by spitting it into multiple smaller components
