+++ src/settings/components/rules/styles.ts
@@ -1,0 +1,44 @@
+export const RULE_STYLES = {
+    INDENT_SIZE: 24,
+    RULE_SPACING: 8,
+    GROUP_PADDING: 16,
+    INPUT_WIDTH: '120px',
+    BADGE_STYLES: {
+        background: 'var(--background-modifier-border)',
+        color: 'var(--text-normal)',
+        borderRadius: '4px',
+        padding: '2px 8px',
+        marginRight: '8px',
+        fontSize: '0.9em'
+    },
+    SELECT_STYLES: `
+        padding: 4px 8px;
+        border-radius: 4px;
+        border: 1px solid var(--background-modifier-border);
+        background: var(--background-primary);
+        color: var(--text-normal);
+        width: 120px;
+    `,
+    CONTAINER_STYLES: `
+        display: flex;
+        flex-direction: column;
+        gap: 8px;
+        padding: 8px;
+        background: var(--background-primary);
+        border-radius: 4px;
+        width: 100%;
+        box-sizing: border-box;
+    `,
+    MAIN_ROW_STYLES: `
+        display: flex;
+        align-items: center;
+        gap: 8px;
+        flex-wrap: wrap;
+    `,
+    BUTTON_CONTAINER_STYLES: `
+        display: flex;
+        gap: 4px;
+        flex: 0 0 auto;
+        margin-left: auto;
+    `
+};
